Merge branch 'master' of github.com:rstudio/rmarkdown into pkgdown-quillt

# Conflicts:
#	.github/workflows/pkgdown.yaml
#	pkgdown/_pkgdown.yml

+++ R/render_site.R
@@ -204,9 +204,9 @@ render_site <- function(input = ".",
 
 #' @rdname render_site
 #' @param preview Whether to list the files to be removed rather than actually
-#'   removing them.
+#'   removing them. Defaulting to TRUE to prevent removing without notice.
 #' @export
-clean_site <- function(input = ".", preview = FALSE, quiet = FALSE,
+clean_site <- function(input = ".", preview = TRUE, quiet = FALSE,
                        encoding = "UTF-8") {
 
   # normalize to a directory
@@ -220,54 +220,82 @@ clean_site <- function(input = ".", preview = FALSE, quiet = FALSE,
   # get the files to be cleaned
   files <- generator$clean()
 
+
+  if (length(files) == 0) {
+    if (preview || !quiet) cat("Nothing to removed. All clean !\n")
+    return(invisible(NULL))
+  }
+
   # if it's just a preview then return the files, otherwise
   # actually remove the files
-  if (preview)
-    files
-  else {
+  if (preview) {
+    cat("These files and folders can probably be removed:\n",
+        paste0("* ", mark_dirs(files)),
+        "\nUse rmarkdown::clean_site(preview = FALSE) to remove them.",
+        sep = "\n")
+  } else {
     if (!quiet) {
-      cat("Removing files: \n")
-      cat(paste0(" ", files), sep = "\n")
+      cat("Removing files: \n",
+          paste0("* ", mark_dirs(files)),
+          sep = "\n")
     }
     unlink(file.path(input, files), recursive = TRUE)
   }
+}
+
+# TODO: Move to xfun - bookdown use it too.
+mark_dirs <- function(files) {
+  i <- file_test("-d", files) & !grepl("/$", files)
+  files[i] <- paste0(files[i], "/")
+  files
 }
 
 #' @rdname render_site
 #' @export
 site_generator <- function(input = ".", output_format = NULL) {
 
+  # look for the closest index file with 'site' metadata
+  root <- tryCatch(
+    proj_root(input, "^index.R?md$", "^\\s*site:.*::.*$"),
+    error = function(e) NULL
+  )
+
   # normalize input
   input <- input_as_dir(input)
 
-  # if we have an index.Rmd (or .md) then check it's yaml for "site:"
-  index <- file.path(input, "index.Rmd")
+  # if none found then look for a _site.yml file
+  if (is.null(root)) {
+    if (file.exists(site_config_file(input))) {
+      return (default_site(input))
+    } else {
+      return(NULL)
+    }
+  }
+
+  # determine the index file (will be index.Rmd or index.md)
+  index <- file.path(root, "index.Rmd")
   if (!file.exists(index))
-    index <- file.path(input, "index.md")
-  if (file.exists(index)) {
+    index <- file.path(root, "index.md")
 
-    # read index.Rmd and extract the front matter
-    front_matter <- yaml_front_matter(index)
+  # is this in a subdir of the site root? (only some generators support this)
+  in_subdir <- !same_path(input, root)
 
-    # is there a custom site generator function?
-    if (!is.null(front_matter$site)) {
+  # read index.Rmd and extract the front matter
+  front_matter <- yaml_front_matter(index)
 
-      create_site_generator <- eval(parse(text = front_matter$site))
-      create_site_generator(input)
+  # create the site generator (passing the root dir)
+  create_site_generator <- eval(parse(text = front_matter$site))
+  generator <- create_site_generator(root)
 
-    # is there a "_site.yml"?
-    } else if (file.exists(site_config_file(input))) {
-
-      default_site(input)
-
-    # no custom site generator or "_site.yml"
+  # if it's in a subdir check to see if the generator supports nested files
+  if (in_subdir) {
+    if (isTRUE(generator$subdirs)) {
+      generator
     } else {
       NULL
     }
-
-  # no index.Rmd or index.md
   } else {
-    NULL
+    generator
   }
 }
 
@@ -614,8 +642,10 @@ input_as_dir <- function(input) {
 
   # ensure the input dir exists
   if (!file.exists(input)) {
-    stop("The specified directory '", normalize_path(input, mustWork = FALSE),
-         "' does not exist.", call. = FALSE)
+    input <- normalize_path(input, mustWork = FALSE)
+    if (!file.exists(input)) stop(
+      "The specified directory '", input, "' does not exist.", call. = FALSE
+    )
   }
 
   # convert from file to directory if necessary
